+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All Notable changes to `viacep-php` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## v2.0.0 - 2019-08-24
+
+### Changed
+- Class `ZipCode` renamed to `ViaCEP`.
+- Method `findByCep` renamed to `findByZipCode`.
+- Method `findByAddress` renamed to `findByStreetName`.
+
+### Removed
+- Method `find` (deprecated) removed.
+
 ## v1.1.0 - 2019-08-24
 
 ### Added